pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coldfront-storage-plugin"
version = "0.0.1"
requires-python = ">=3.10"
authors = [
    { name = "Chris Barnett" }
]
description = "Coldfront Storage Plugin"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["high-performance-computing", "resource-allocation"]
classifiers = [
    'Programming Language :: Python :: 3',
    'Framework :: Django :: 6.0',
    'Topic :: Scientific/Engineering',
    'Topic :: System :: Systems Administration',
    'Topic :: Internet :: WWW/HTTP :: WSGI :: Application',
]
dependencies = [
    "coldfront",
    "coldfront_utils@git+https://github.com/Tufts-Technology-Services/coldfront-utils.git",
]

[project.urls]
"Source Code" = "https://github.com/Tufts-Technology-Services/coldfront-storage-plugin"

[project.optional-dependencies]
starfish = [
    "starfish_api_client@git+https://github.com/Tufts-Technology-Services/starfish-api-client.git",
]

vast = [
    "vast_api_client@https://github.com/Tufts-Technology-Services/vast-api-client.git",
]

truenas = [
    "truenas_utils@https://github.com/Tufts-Technology-Services/truenas-utils.git",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "reuse>=5.0.2",
    "ruff>=0.11.4",
]

[tool.hatch.build.targets.wheel]
packages = ["storage"]

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "S308"]

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true
